move ResetForTest to be available only during tests
