+++ internal/global/state.go
@@ -18,7 +18,6 @@ import (
 	"errors"
 	"sync"
 	"sync/atomic"
-	"testing"
 
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/trace"
@@ -113,15 +112,4 @@ func defaultPropagatorsValue() *atomic.Value {
 	v := &atomic.Value{}
 	v.Store(propagatorsHolder{tm: newTextMapPropagator()})
 	return v
-}
-
-// ResetForTest configures the test to restores the initial global state during
-// its Cleanup step
-func ResetForTest(t testing.TB) {
-	t.Cleanup(func() {
-		globalTracer = defaultTracerValue()
-		globalPropagators = defaultPropagatorsValue()
-		delegateTraceOnce = sync.Once{}
-		delegateTextMapPropagatorOnce = sync.Once{}
-	})
 }
