+++ internal/global/util_test.go
@@ -1,0 +1,31 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package global
+
+import (
+	"sync"
+	"testing"
+)
+
+// ResetForTest configures the test to restores the initial global state during
+// its Cleanup step
+func ResetForTest(t testing.TB) {
+	t.Cleanup(func() {
+		globalTracer = defaultTracerValue()
+		globalPropagators = defaultPropagatorsValue()
+		delegateTraceOnce = sync.Once{}
+		delegateTextMapPropagatorOnce = sync.Once{}
+	})
+}
